requirements.txt
Django==5.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.1.1
pillow==12.0.0
djangorestframework_simplejwt==5.5.1




